requirements.txt
configparser==5.3.0
discord.py==2.0.1
